Centralise l'initialisation de Celery
