+++ worker.py
@@ -1,27 +1,6 @@
-import os
-from celery import Celery
-
-from app import app
+from celery_app import celery
 from observability.logging import get_logger
 from celery import signals
-
-broker = os.getenv("CELERY_BROKER_URL", "redis://localhost:6379/0")
-backend = os.getenv("CELERY_BACKEND_URL", broker)
-
-celery = Celery(
-    "cadlytics", broker=broker, backend=backend, include=["tasks.conversion", "tasks.dfm"]
-)
-
-celery.conf.update(task_track_started=True, result_expires=3600)
-
-
-class ContextTask(celery.Task):
-    def __call__(self, *args, **kwargs):
-        with app.app_context():
-            return super().__call__(*args, **kwargs)
-
-
-celery.Task = ContextTask
 
 logger = get_logger(__name__)
